+++ quant_trader/strategy/template/single_time_frame_strategy.h
@@ -31,7 +31,7 @@ public:
     void setBarList(const QMap<int, QPair<QList<Bar>*, Bar*>> &listAndLast) override;
 
     void checkIfNewBar(int newBarTimeFrame) override;
-    void onNewTick(int time, double lastPrice) override;
+    void onNewTick(qint64 time, double lastPrice) override;
 
 };
